Fix robocap deps, remove deprecated egoexo converters, monoprior viewer/dtype fixes
robocap-slam:
- Add pytorch-cpu to the robocap feature so the shared simplecv.video_io
  top-level `import torch` resolves. robocap uses the cv2 VideoReader and never
  executes torch; pytorch-gpu would conflict with robocap's cuda-13 toolchain
  (the shared cuda feature is cuda 12.9, linux-64 only) and its aarch64 target.

egoexo-forge:
- Remove the deferred, unused SAM3D-body-parquet feature in full: converter,
  mesh/blueprint helpers, the sam3d-body-rerun mhr70 shim, its tests, the CLI
  tool, and the pixi task. (The standalone sam3d-body-rerun package is untouched.)
- Replace batch_convert with a deprecation stub pointing at the relocated
  simplecv.apis.batch_raw_to_rrd; update the task description and README.
- Drop the obsolete batch_ingest_rrd stub module.

monoprior:
- Cast simplecv's float64 Intrinsics.k_matrix to float32 at pipeline boundaries
  (Float annotations + explicit .astype) so beartype stops rejecting the reads.
- Log raw rr.Image instead of .compress() to avoid the Rerun 0.33 native-viewer
  EncodedImage deadlock (verbose calibration path).
- Extract the shared parse_preprocessing_mode helper for the VGGT Gradio UIs.

deps / tooling:
- Pin mkl >=2025.3,<2026 in the mast3r-slam/faiss feature (faiss links the
  .so.2 soname; mkl 2026 ships .so.3 -> ImportError).
- Bump pyrefly >=1.1.0 and refresh the simplecv pyrefly baseline.

Other: pyvrs-viewer neutral-chroma byte caching, wilor-nano config dataclass
simplification, prompt-da resize inlining, and assorted doc updates.

+++ packages/monoprior/monopriors/apis/multiview_inference.py
@@ -424,7 +424,8 @@ def mv_pred_to_pointcloud(
     ).astype(np.float32)
     K_list: list[Float32[ndarray, "3 3"]] = []
     for mv_pred in mv_pred_list:
-        K_33: Float32[ndarray, "3 3"] | None = mv_pred.pinhole_param.intrinsics.k_matrix
+        # simplecv's Intrinsics stores k_matrix as float64; cast to float32 for the pipeline
+        K_33: Float[ndarray, "3 3"] | None = mv_pred.pinhole_param.intrinsics.k_matrix
         if K_33 is None:
             raise ValueError("VGGT prediction must include camera intrinsics.")
         K_list.append(K_33.astype(np.float32))
@@ -570,10 +571,11 @@ def run_inference(config: VGGTInferenceConfig) -> None:
             image_plane_distance=0.1,
             static=True,
         )
-        K_33: Float32[ndarray, "3 3"] | None = mv_pred.pinhole_param.intrinsics.k_matrix
+        # simplecv's Intrinsics stores k_matrix as float64; cast to float32 for the stack
+        K_33: Float[ndarray, "3 3"] | None = mv_pred.pinhole_param.intrinsics.k_matrix
         if K_33 is None:
             raise ValueError("COLMAP export requires camera intrinsics.")
-        intri_stack_list.append(K_33)
+        intri_stack_list.append(K_33.astype(np.float32))
 
         rr.log(
             f"{pinhole_log_path}/image",
